standardize runPromise inside namespace, flatten Account/Truncate facades
- FileWatcher, Vcs, Installation: move runPromise inside namespace
- Truncate: flatten truncate-effect.ts → truncate.ts
- Account: flatten effect.ts → index.ts
- Server: use Vcs.branch() facade instead of raw runPromise
- Auth stays split (dual Schema/Zod types)
