+++ packages/opencode/src/tool/question.ts
@@ -1,7 +1,6 @@
 import z from "zod"
 import { Tool } from "./tool"
 import { Question } from "../question"
-import { Question as QuestionApi } from "../question"
 import DESCRIPTION from "./question.txt"
 
 export const QuestionTool = Tool.define("question", {
@@ -10,7 +9,7 @@ export const QuestionTool = Tool.define("question", {
     questions: z.array(Question.Info.omit({ custom: true })).describe("Questions to ask"),
   }),
   async execute(params, ctx) {
-    const answers = await QuestionApi.ask({
+    const answers = await Question.ask({
       sessionID: ctx.sessionID,
       questions: params.questions,
       tool: ctx.callID ? { messageID: ctx.messageID, callID: ctx.callID } : undefined,
